refactor: change file structure in API
