+++ backend/api/controllers/auth/logoutController.js
@@ -1,0 +1,17 @@
+const express = require('express');
+const session = require('express-session');
+
+
+const logout = async (req, res) => {
+    req.session.destroy(err => {
+        if (err) {
+          return res.sendStatus(500);
+        }
+        res.clearCookie('connect.sid');
+        res.sendStatus(200);
+      });
+};
+
+module.exports = {
+  logout,
+};
