+++ backend/api/package.json
@@ -13,6 +13,7 @@
     "bcrypt": "^5.1.1",
     "connect-mongo": "^5.1.0",
     "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
     "express": "^4.21.0",
     "express-session": "^1.18.1",
     "form-data": "^4.0.1",
