+++ backend/api/services/mailService.js
@@ -1,3 +1,4 @@
+require('dotenv').config();
 const formData = require('form-data');
 const Mailgun = require('mailgun.js');
 const mailgun = new Mailgun(formData);
